Use a separate assembly, pack as tool
