+++ InsertionChangeLogGenerator/InsertionChangeLogGenerator.sln
@@ -1,0 +1,34 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.6.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InsertionChangeLogGenerator", "InsertionChangeLogGenerator.csproj", "{5099C360-767C-4108-A204-8BB847545E05}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5099C360-767C-4108-A204-8BB847545E05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5099C360-767C-4108-A204-8BB847545E05}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5099C360-767C-4108-A204-8BB847545E05}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{5099C360-767C-4108-A204-8BB847545E05}.Debug|x64.Build.0 = Debug|Any CPU
+		{5099C360-767C-4108-A204-8BB847545E05}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5099C360-767C-4108-A204-8BB847545E05}.Debug|x86.Build.0 = Debug|Any CPU
+		{5099C360-767C-4108-A204-8BB847545E05}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5099C360-767C-4108-A204-8BB847545E05}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5099C360-767C-4108-A204-8BB847545E05}.Release|x64.ActiveCfg = Release|Any CPU
+		{5099C360-767C-4108-A204-8BB847545E05}.Release|x64.Build.0 = Release|Any CPU
+		{5099C360-767C-4108-A204-8BB847545E05}.Release|x86.ActiveCfg = Release|Any CPU
+		{5099C360-767C-4108-A204-8BB847545E05}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
